Multiplication result converted to larger type
CWE-190
CWE-192
CWE-197
CWE-681

Use wider arithmetic at multiplication time by casting one operand before
multiplying, then divide in the wider type, and cast back to uint32 for
assignment.

Best fix in src/UploadDiskIOThread.cpp at line 550 region:

- Change uint32 payloadSize = nPacketSize*oldSize/newsize;
- To: uint32 payloadSize = static_cast<uint32>((static_cast<uint64>(nPacketSize) * oldSize) / newsize);

This ensures multiplication cannot overflow 32-bit before division.

No import/dependency changes are needed.

+++ src/UploadDiskIOThread.cpp
@@ -547,7 +547,9 @@ void CUploadDiskIOThread::CreatePackedPackets(const uint8_t* buffer, uint64 star
 		CPacket* packet = new CPacket(data, OP_EMULEPROT, (isLargeBlock ? OP_COMPRESSEDPART_I64 : OP_COMPRESSEDPART));
 
 		// approximate payload size
-		uint32 payloadSize = nPacketSize*oldSize/newsize;
+		uint32 payloadSize = static_cast<uint32>(
+			(static_cast<uint64>(nPacketSize) * oldSize) /
+			newsize);
 
 		if (togo == 0 && totalPayloadSize+payloadSize < oldSize) {
 			payloadSize = oldSize-totalPayloadSize;
